Refactor: rename NumWordConverter to num_word_converter
This change standardizes the naming convention across the project to use snake_case instead of CamelCase. It includes updates to module imports, package names, and test references to ensure consistency.
