--- NumWordConverter/errors/__init__.py
@@ -1,2 +1,0 @@
-from NumWordConverter.errors.word2num_errors.errors import *
-from NumWordConverter.errors.num2word_errors.errors import *
