+++ num_word_converter/errors/__init__.py
@@ -1,0 +1,2 @@
+from num_word_converter.errors.word2num_errors.errors import *
+from num_word_converter.errors.num2word_errors.errors import *
