+++ setup.py
@@ -1,7 +1,7 @@
 from setuptools import setup, find_packages
 
 setup(
-    name="NumWordConverter",
+    name="num_word_converter",
     version="0.1",
     packages=find_packages(),
